feat(model,view): Capture an archi
User can now click on a delegate to mark it as captured.

The data representation have changed. Qt JSON is meant to be use read-
only so we can't use it to store who is captured.

Now ArchiTrackteur will read the JSON and convert it to a c++ structure.
Only the captured state will be saved.