feat(model): Captured archi can be saved and loaded
All archi captured are now saved when the user quit the application.
They are automaticaly loaded when he open the app again.

The captured archi are stored in a map<int,int>. The key is the ID of
the archi and the value is the number of archi the user currently have.